Redefine is_prime_* interface
is_prime functions now return a truth value depending on wether it is
decided to be a prime or not. To satisfy the previous functionaly a new
function, h_prime_get_index is introduced. It returns the index of that
prime if it is found in the array or -1 if it is not found.